requirements.txt
langchain-core
langchain-openai
langchain-anthropic
langgraph
yfinance
python-dotenv
scikit-learn
pandas
numpy
matplotlib
seaborn
plotly
joblib